# api-docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/ordens-servicos": {
      "get": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Lista todas as ordens de serviço de forma paginada",
        "operationId": "listar_1",
        "parameters": [
          {
            "name": "pagina",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 0
            }
          },
          {
            "name": "quantidade",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ordens de serviço retornada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Abre uma nova ordem de serviço",
        "operationId": "criar",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CriarOrdemDeServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Ordem de serviço aberta com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/ordens-servicos/{id}/finalizacoes": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Finaliza a execução de uma ordem de serviço",
        "operationId": "finalizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OS finalizada com sucesso"
          },
          "404": {
            "description": "Ordem de serviço não encontrada"
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}/finalizacoes",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/ordens-servicos/{id}/execucoes": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Inicia a execução de uma ordem de serviço após aprovação do orçamento",
        "operationId": "executar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecutarOrdemServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Execução da OS iniciada com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ordem de serviço não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "409": {
            "description": "Orçamento não aprovado ou OS em estado incorreto",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}/execucoes",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/ordens-servicos/{id}/entregas": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Registra a entrega do veículo ao cliente",
        "operationId": "entregar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Veículo entregue com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ordem de serviço não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}/entregas",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
    "/ordens-servicos/{id}/diagnosticos/finalizacoes": {
      "post": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Finaliza o diagnóstico de uma OS, informando peças e serviços necessários",
        "operationId": "finalizarDiagnostico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizarDiagnosticoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Diagnóstico finalizado e orçamento gerado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ordem de serviço não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}/diagnosticos/finalizacoes",
          "httpMethod": "POST",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/servicos/{id}": {
      "get": {
        "tags": [
          "Serviços"
        ],
        "summary": "Buscar serviço por ID",
        "description": "Retorna os detalhes de um serviço específico",
        "operationId": "buscar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do serviço",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Serviço encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Serviço não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos/{id}",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "delete": {
        "tags": [
          "Serviços"
        ],
        "summary": "Deletar serviço",
        "description": "Remove um serviço pelo seu ID",
        "operationId": "deletar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do serviço",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Serviço deletado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Serviço não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos/{id}",
          "httpMethod": "DELETE",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      },
      "patch": {
        "tags": [
          "Serviços"
        ],
        "summary": "Atualizar serviço",
        "description": "Atualiza os dados de um serviço existente",
        "operationId": "atualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do serviço",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtualizarServicoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Serviço atualizado com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Serviço não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/servicos/{id}",
          "httpMethod": "PATCH",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
    "/ordens-servicos/{id}": {
      "get": {
        "tags": [
          "Ordens de Serviço"
        ],
        "summary": "Busca uma ordem de serviço por ID",
        "operationId": "buscar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ordem de serviço encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Ordem de serviço não encontrada",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "uri": "API_DNS/ordens-servicos/{id}",
          "httpMethod": "GET",
          "type": "http_proxy",
          "passthroughBehavior": "when_no_match",
          "connectionType": "INTERNET"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CriarOrdemDeServicoRequest": {
        "type": "object",
        "properties": {
          "veiculoId": {
            "type": "string",
            "format": "uuid"
          },
          "clienteId": {
            "type": "string",
            "format": "uuid"
          },
          "pecasInsumos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemPecaInsumoRequest"
            }
          },
          "servicos": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "clienteId",
          "veiculoId"
        ]
      },
      "ItemPecaInsumoRequest": {
        "type": "object",
        "properties": {
          "idPecaInsumo": {
            "type": "string",
            "format": "uuid"
          },
          "qtd": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ExecutarOrdemServicoRequest": {
        "type": "object",
        "properties": {
          "orcamentoId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "FinalizarDiagnosticoRequest": {
        "type": "object",
        "properties": {
          "idPecasNecessarias": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemPeca"
            }
          },
          "idServicosNecessarios": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "ItemPeca": {
        "type": "object",
        "properties": {
          "idPeca": {
            "type": "string",
            "format": "uuid"
          },
          "qtd": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "AtualizarServicoRequest": {
        "type": "object",
        "description": "Dados para atualização do serviço",
        "properties": {
          "preco": {
            "type": "number"
          },
          "ativo": {
            "type": "boolean"
          }
        },
        "required": [
          "ativo",
          "preco"
        ]
      },
...
    }
  }
}